Move notificationDrivers map to NotificationSyncer
